first draft submission

+++ controllers/clothingItems.js
@@ -1,5 +1,11 @@
 const clothingItem = require("../models/clothingItem");
-const { SERVER_ERROR, BAD_REQUEST, NOT_FOUND } = require("../utils/errors");
+const {
+  SERVER_ERROR,
+  BAD_REQUEST,
+  NOT_FOUND,
+  UNAUTHORIZED_REQUEST,
+  FORBIDDEN_REQUEST,
+} = require("../utils/errors");
 
 const getClothingItems = (req, res) => {
   clothingItem
@@ -15,7 +21,7 @@ const getClothingItems = (req, res) => {
 
 const addClothingItem = (req, res) => {
   const { name, weather, imageUrl } = req.body;
-  const { _id } = req.user._id;
+  const _id = req.user._id;
   clothingItem
     .create({ name, weather, imageUrl, owner: _id })
     .then((item) => {
@@ -32,18 +38,36 @@ const addClothingItem = (req, res) => {
 
 const deleteClothingItem = (req, res) => {
   const { id } = req.params;
+
   clothingItem
-    .findByIdAndDelete(id)
+    .findById(id)
     .orFail()
     .then((item) => {
-      res.status(200).send(item);
+      if (item.owner.toString() === req.user._id.toString()) {
+        clothingItem
+          .findByIdAndDelete(id)
+          .orFail()
+          .then((item) => {
+            res.status(200).send(item);
+          })
+          .catch((err) => {
+            if (err.name === "DocumentNotFoundError") {
+              return NOT_FOUND(err, res);
+            }
+            if (err.name === "CastError") {
+              return BAD_REQUEST(err, res);
+            }
+            return SERVER_ERROR(err, res);
+          });
+      } else {
+        return FORBIDDEN_REQUEST({}, res);
+      }
     })
     .catch((err) => {
-      console.error(err);
       if (err.name === "DocumentNotFoundError") {
         return NOT_FOUND(err, res);
       }
-      if (err.name === "CastError" || err.name === "ValidationError") {
+      if (err.name === "CastError") {
         return BAD_REQUEST(err, res);
       }
       return SERVER_ERROR(err, res);
